поделил функционал логики на 3 сервиса
